Split SeedbankTree class into SeedbankTree and SeedbankTreeInitialiser
Two classes mimic the structure of MultiTypeTree and StructuredCoalescentMultiTypeTree classes from the MultiTypeTree package. This way, the secondary initialiser class implements the StateNodeInitialiser interface and acts as an initialiser for a given, separate  m_initial input tree. SeedbankTreeInitialiser differs from its MultiTypeTree counterpart in that it requires the m_initial input tree, whereas the SCMTT class acts as the initialised tree itself.